response to second @mikesposito review
